+++ environment.d.ts
@@ -19,6 +19,7 @@ declare global {
       DATABASE_URI: string
       TEST_DATABASE_URI: string
       ENCRYPT_KEY: string
+      SEVENTV_GQL_TOKEN: string
     }
   }
 }
@@ -39,4 +40,3 @@ export interface Environment {
   MAIN_USER_ID: string
   TEST_DATABASE_URI: string
 }
-