+++ src/commands/export-commands.ts
@@ -20,10 +20,12 @@ import { RemoveSuggestCommand } from '@src/commands/cmd/rmsuggestion'
 import { SuggestCommand } from '@src/commands/cmd/suggest'
 import { WebsiteCommand } from '@src/commands/cmd/website'
 import { Command } from '@src/commands/types'
+import { AddCommand } from '@src/commands/cmd/add'
+import { RemovemeCommand } from '@src/commands/cmd/removeme'
 
 const commands: Command[] = [
   new PingCommand(),
-  new RemoveCommand(),
+  new RemovemeCommand(),
   new GithubCommand(),
   new WebsiteCommand(),
   new NotifyCommand(),
@@ -41,7 +43,9 @@ const commands: Command[] = [
   new EvalCommand(),
   new ColorHistoryCommand(),
   new SetLevelCommand(),
-  new TimeoutsCommand()
+  new TimeoutsCommand(),
+  new AddCommand(),
+  new RemoveCommand()
 ]
 
 export default commands