+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - command to set level of user
 - command to update status of suggestion
 - save current status of suggestions
+- adding 7tv emotes
+- removing 7tv emotes
 
 ### Changed 
 - Removing repo from global db object